does some required changes and add some extra option

+++ client/src/components/admin/NewsTable/NewsTable.jsx
@@ -23,6 +23,8 @@ const NewsTable = () => {
     fetchProperties();
 }, []);
 const handleDelete=async(id)=>{
+  const confirmDelete = window.confirm("Are you sure you want to delete this message?");
+  if (!confirmDelete) return;
     try {
       const response = await fetch(`/api/news/delete/${id}`, {
           method: "DELETE",
