+++ client/src/components/admin/Users/TableUser.jsx
@@ -22,6 +22,8 @@ const TableUser = () => {
 }, []);
 
 const handleDelete=async(id)=>{
+  const confirmDelete = window.confirm("Are you sure you want to delete this message?");
+  if (!confirmDelete) return;
   try {
     const response = await fetch(`/api/user/deleteuser/${id}`, {
         method: "DELETE",
